transcription-summarize/requirements.txt
h2o-wave==0.26.2
openai-whisper==20230918
pandas==2.1.1
glob2==0.7
gradio_client==0.6.1
loguru==0.6.0